Modify GitLoaderContext to load any assembly in the folder
When the entry-point task is loaded in an AssemblyLoadContext (as it is
in MSBuild 16.5+ on .NET Core), the default ALC cannot load assemblies
from next to the task location, so assemblies that are redistributed with
the GitVersionTask package aren't automatically found. Extend the task's
private ALC to support loading any assembly from the task's folder.