Prevent GenerateGitVersionInformation in XAML generated projects
During XAML'S build magic a temporary project may get generated and
build (target "GenerateTemporaryTargetAssembly"). The generated project
file already contains compile-items with the generated files from the
original project, this `GitVerstionInformation.g.cs` is listed. When the
GitVersion task now runs again, the file is added another time resulting
in an error of CSC.
This patch tries to detect this situation and prevent the execution of
GenerateGitVersionInformation by checking for a special property
injected by GenerateTemporaryTargetAssembly. Unfortunately this is not
documented and thus may change.

See also
- https://referencesource.microsoft.com/#PresentationBuildTasks/BuildTasks/Microsoft/Build/Tasks/Windows/GenerateTemporaryTargetAssembly.cs,163
- #2228