Update SDK version and link global.json to solution
Updated the SDK version in global.json from 7.0.103 to 7.0.306 to stay up-to-date with the latest version and potentially use the new features or fixes introduced in it. Also, linked the global.json file to the solution (GitVersion.sln) for easier access and more structured file organization.

Update Microsoft.CodeAnalysis.CSharp package version

The package version for "Microsoft.CodeAnalysis.CSharp" was upgraded from version 4.4.0 to 4.6.0.